+++ src/core/core.hpp
